pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["yuxiaofei <531189371@qq.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
sqlalchemy = "^1.4.0b"
python-dotenv = "^0.15.0"
uvicorn = "^0.13.4"
python-jose = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "^20.8b1"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"
httpx = "^0.16.1"
requests = "^2.25.1"
alembic = "^1.5.5"
typer = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true

[tool.poetry.scripts]
todolist = "src.cli:main"
